FIX #71 - Compiler warnings from modern g++
Fix a number of compiler warnings related to use of ... with format
strings that don't (or might not) match the supplied parameters.

Most were most likely harmless but a couple looked like they would
cause unintended results.

g++ does not like use of printf etc with a variable for the format
string parameter (because it will behave unpredicatably if the
string passed to it contains a %). I don't like using appendf with a
single constant string parameter (because it's inefficient).

+++ deployment/deployutils/configenvhelper.cpp
@@ -1270,7 +1270,7 @@ bool CConfigEnvHelper::AddNewNodes(IPropertyTree* pThor, const char* szType, int
 
     if (validate)
     {
-        for (size_t i = 0; i < computers.size(); i++)
+        for (int i = 0; i < (int) computers.size(); i++)
             CheckTopologyComputerUse(computers[i], pThor, usageList);
     }
 
@@ -1278,7 +1278,7 @@ bool CConfigEnvHelper::AddNewNodes(IPropertyTree* pThor, const char* szType, int
         return false;
 
     // Iterate through computer list
-    for (size_t i = 0; i < computers.size(); i++)
+    for (int i = 0; i < (int) computers.size(); i++)
     {
         // Check if we can add this computer
         if (skipExisting && !CheckTopologyComputerUse(computers[i], pThor, usageList))
